+++ fastly/fastly.go
@@ -5,6 +5,19 @@ import (
 	"encoding"
 )
 
+type BatchOperation string
+
+const (
+	CreateBatchOperation BatchOperation = "create"
+	UpdateBatchOperation BatchOperation = "update"
+	UpsertBatchOperation BatchOperation = "upsert"
+	DeleteBatchOperation BatchOperation = "delete"
+
+	// Represents the maximum number of operations that can be sent within a single batch request.
+	// This is currently not documented in the API.
+	BatchModifyMaximumOperations = 1000
+)
+
 type statusResp struct {
 	Status string
 	Msg    string